+++ framework/model/src/main/java/is/codion/framework/model/AbstractEntityTableModel.java
@@ -43,9 +43,6 @@ import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 
-import static is.codion.framework.db.EntityConnection.Select.where;
-import static is.codion.framework.domain.entity.condition.Condition.keys;
-import static java.util.Collections.emptyList;
 import static java.util.Collections.singleton;
 import static java.util.Objects.requireNonNull;
 import static java.util.function.Function.identity;
@@ -166,11 +163,6 @@ public abstract class AbstractEntityTableModel<M extends EntityModel<M, E, T, R>
 		selection().items().set(new SelectByKeyPredicate(requireNonNull(keys)));
 	}
 
-	@Override
-	public final RefreshTask refreshTask(Collection<Entity.Key> keys) {
-		return new DefaultRefreshTask(requireNonNull(keys));
-	}
-
 	/**
 	 * @return the underlying model
 	 */
@@ -287,42 +279,6 @@ public abstract class AbstractEntityTableModel<M extends EntityModel<M, E, T, R>
 		private void updated(EntityType entityType, Map<Entity.Key, Entity> entities) {
 			entityDefinition().foreignKeys().get(entityType)
 							.forEach(foreignKey -> AbstractEntityTableModel.this.updated(foreignKey, entities));
-		}
-	}
-
-	private final class DefaultRefreshTask implements RefreshTask {
-
-		private final Collection<Entity.Key> keys;
-
-		private DefaultRefreshTask(Collection<Entity.Key> keys) {
-			this.keys = keys;
-		}
-
-		@Override
-		public Result perform() {
-			if (keys.isEmpty()) {
-				return new DefaultRefreshResult(emptyList());
-			}
-
-			return new DefaultRefreshResult(connection().select(where(keys(keys))
-							.attributes(queryModel.attributes().defaults().get())
-							.include(queryModel.attributes().include().get())
-							.exclude(queryModel.attributes().exclude().get())
-							.build()));
-		}
-	}
-
-	private final class DefaultRefreshResult implements RefreshTask.Result {
-
-		private final Collection<Entity> entities;
-
-		private DefaultRefreshResult(Collection<Entity> entities) {
-			this.entities = entities;
-		}
-
-		@Override
-		public void handle() {
-			replace(entities);
 		}
 	}